requirements.txt
openai
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
mysqlclient
